Converting the proxy version of directlyOn(...) to @Direct with reflector(...).
This change is a continuation of the migration towards Reflector. It was preceded by the conversion of the String version of directlyOn(...). Based on testing, this change should improve performance by another 4 or 5% for the top 25% of tests. The main problem with the proxy version of directlyOn(...) was the overhead of class definition and instance creation. Proxy classes had to be constantly regenerated by ASM because they are not cached across tests and instantiating new proxy instances for large classes (Activity, View, TextView, Resources, etc.) is slow.

PiperOrigin-RevId: 384243997